backend/requirements.txt
# FastAPI and Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0

# Google Cloud Vision (OCR)
google-cloud-vision==3.4.5

# Web Push Notifications
pywebpush==1.14.0
py-vapid==1.9.0

# Image Processing
Pillow==10.1.0

# HTTP Client
httpx==0.25.1
aiofiles==23.2.1

# Task Scheduling
apscheduler==3.10.4

# CSV Export
pandas==2.1.3

# Utilities
python-dateutil==2.8.2
pytz==2023.3
